+++ packages/grafana-data/src/dataframe/index.ts
@@ -2,3 +2,4 @@ export * from './DataFrameView';
 export * from './FieldCache';
 export * from './CircularDataFrame';
 export * from './MutableDataFrame';
+export * from './processDataFrame';
