+++ packages/grafana-data/src/transformations/matchers/fieldTypeMatcher.test.ts
@@ -1,7 +1,7 @@
 import { FieldType } from '../../types/dataFrame';
 import { fieldMatchers } from './matchers';
 import { FieldMatcherID } from './ids';
-import { toDataFrame } from '../../utils/processDataFrame';
+import { toDataFrame } from '../../dataframe/processDataFrame';
 
 export const simpleSeriesWithTypes = toDataFrame({
   fields: [
