+++ packages/grafana-data/src/transformations/transformers/append.test.ts
@@ -1,5 +1,5 @@
 import { DataTransformerID } from './ids';
-import { toDataFrame } from '../../utils/processDataFrame';
+import { toDataFrame } from '../../dataframe/processDataFrame';
 import { transformDataFrame } from '../transformDataFrame';
 import { transformersRegistry } from '../transformersRegistry';
