+++ packages/grafana-data/src/utils/fieldReducer.test.ts
@@ -3,7 +3,7 @@ import difference from 'lodash/difference';
 import { fieldReducers, ReducerID, reduceField } from './fieldReducer';
 
 import { Field, FieldType } from '../types/index';
-import { guessFieldTypeFromValue } from './processDataFrame';
+import { guessFieldTypeFromValue } from '../dataframe/processDataFrame';
 import { MutableDataFrame } from '../dataframe/MutableDataFrame';
 import { ArrayVector } from '../vector/ArrayVector';
